Refactor: 유저관련 기능 코드 수정

+++ grid/src/main/java/org/highfives/grid/user/repository/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<Employee, Long> {
+public interface UserRepository extends JpaRepository<Employee, Integer> {
 }
